# OpenApi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SimpleCare.API",
    "version": "1.0"
  },
  "paths": {
    "/api/bed-ward/patients": {
      "get": {
        "tags": [
          "BedWard"
        ],
        "operationId": "GetBedWardPatients",
        "parameters": [
          {
            "name": "wardId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BedWardPatientListItem"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/bed-ward/patients/{patientId}": {
      "get": {
        "tags": [
          "BedWard"
        ],
        "operationId": "GetBedWardPatient",
        "parameters": [
          {
            "name": "patientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BedWardPatient"
                }
              }
            }
          }
        }
      }
    },
    "/api/bed-ward/patients/admit": {
      "post": {
        "tags": [
          "BedWard"
        ],
        "operationId": "AdmitPatient",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientAdmission"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientAdmission"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PatientAdmission"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/emergency-wards/patients": {
      "get": {
        "tags": [
          "EmergencyWard"
        ],
        "operationId": "GetEmergencyPatients",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/EmergencyPatientStatus"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EmergencyPatientListItem"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/emergency-wards/patients/{patientId}": {
      "get": {
        "tags": [
          "EmergencyWard"
        ],
        "operationId": "GetEmergencyPatient",
        "parameters": [
          {
            "name": "patientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmergencyPatient"
                }
              }
            }
          }
        }
      }
    },
    "/api/emergency-wards/recipient-wards": {
      "get": {
        "tags": [
          "EmergencyWard"
        ],
        "operationId": "GetRecipientWards",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecipientWardItem"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/emergency-wards/patients/register": {
      "post": {
        "tags": [
          "EmergencyWard"
        ],
        "operationId": "RegisterEmergencyPatient",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmergencyRegistration"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EmergencyRegistration"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EmergencyRegistration"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/emergency-wards/patients/transfer": {
      "post": {
        "tags": [
          "EmergencyWard"
        ],
        "operationId": "TransferEmergencyPatient",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TransferRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/orderly/tasks": {
      "get": {
        "tags": [
          "Orderly"
        ],
        "operationId": "GetOrderlyTasks",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrderlyTaskListItem"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/orderly/Tasks/start": {
      "post": {
        "tags": [
          "Orderly"
        ],
        "operationId": "StartTask",
        "parameters": [
          {
            "name": "taskId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/orderly/Tasks/end": {
      "post": {
        "tags": [
          "Orderly"
        ],
        "operationId": "EndTask",
        "parameters": [
          {
            "name": "taskId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BedWardPatient": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "personalIdentifier": {
            "type": "string",
            "nullable": true
          },
          "familyName": {
            "type": "string",
            "nullable": true
          },
          "givenNames": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "BedWardPatientListItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "personalIdentifier": {
            "type": "string",
            "nullable": true
          },
          "familyName": {
            "type": "string",
            "nullable": true
          },
          "givenNames": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "EmergencyPatient": {
        "type": "object",
        "properties": {
          "patientId": {
            "type": "string",
            "format": "uuid"
          },
          "familyName": {
            "type": "string",
            "nullable": true
          },
          "givenNames": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "EmergencyPatientListItem": {
        "type": "object",
        "properties": {
          "patientId": {
            "type": "string",
            "format": "uuid"
          },
          "familyName": {
            "type": "string",
            "nullable": true
          },
          "givenNames": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/EmergencyPatientStatus"
          }
        },
        "additionalProperties": false
      },
      "EmergencyPatientStatus": {
        "enum": [
          "Registered",
          "InTransfer",
          "Discharged"
        ],
        "type": "string"
      },
      "EmergencyRegistration": {
        "type": "object",
        "properties": {
          "personalIdentifier": {
            "type": "string",
            "nullable": true
          },
          "familyName": {
            "type": "string",
            "nullable": true
          },
          "givenNames": {
            "type": "string",
            "nullable": true
          },
          "reason": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "OrderlyTaskListItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "startLocation": {
            "type": "string",
            "nullable": true
          },
          "endLocation": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "startTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "endTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PatientAdmission": {
        "type": "object",
        "properties": {
          "patientId": {
            "type": "string",
            "format": "uuid"
          },
          "wardId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "RecipientWardItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "identifier": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TransferRequest": {
        "type": "object",
        "properties": {
          "patientId": {
            "type": "string",
            "format": "uuid"
          },
          "familyName": {
            "type": "string",
            "nullable": true
          },
          "givenNames": {
            "type": "string",
            "nullable": true
          },
          "wardId": {
            "type": "string",
            "format": "uuid"
          },
          "reason": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}